remove duplicate title welcome page
