+++ src/other/pages/welcome/index.scss
@@ -1,0 +1,40 @@
+@import '../utils/accessory-story.scss';
+@import '../utils/canvas-content-padding.scss';
+
+@include canvas-padding;
+@include content-padding;
+
+@include accessory-story-content;
+@include accessory-story-wrapper;
+
+.welcome {
+    &__commercial-hero-banner-button-wrapper {
+        display: flex;
+    }
+
+    &__commercial-hero-banner-image {
+        display: none;
+        @include breakpoint($mediaquery-md-up) {
+            display: block;
+            position: absolute;
+            top: 345px;
+            right: 30px;
+            height: 360px;
+        }
+        @include breakpoint($mediaquery-xl-up) {
+            right: 6%;
+            top: 325px;
+            height: 420px;
+        }
+    }
+
+    &__text-with-link {
+        @include text(t2);
+        max-width: 750px;
+    }
+
+    &__response-date {
+        // The important flag is needed, because the component iself sets its styles to the host element internally.
+        margin-bottom: 20px !important;
+    }
+}
