+++ src/other/pages/welcome/story.js
@@ -4,6 +4,29 @@ import '../../../components/10-atoms/text';
 import '../utils/contact-footer';
 import styles from './index.scss';
 
+const getFormattedGitCommitMessage = answerJson => {
+  const keyword = 'Publish\n\n - ';
+  const seperator = '<br>';
+
+  if (
+    answerJson.items.length > 0 &&
+    answerJson.items[0].commit.message.startsWith(keyword)
+  ) {
+    const formattedMessage = answerJson.items[0].commit.message
+      .replace(keyword, '')
+      .replace(/ - /g, seperator);
+
+    return formattedMessage;
+  }
+
+  return '';
+};
+
+const getDateFromGitCommit = answerJson => {
+  const formattedDate = new Date(answerJson.items[0].commit.author.date);
+  return formattedDate.toLocaleString(navigator.language);
+};
+
 export default {
   title: 'Welcome',
   decorators: [],
@@ -49,7 +72,7 @@ export const ToPatternLibrary = () => {
           concepts
         </p>
         <div
-          class="landingpage__commercial-hero-banner-button-wrapper"
+          class="welcome__commercial-hero-banner-button-wrapper"
           slot="button"
         >
           <axa-button-link
@@ -68,7 +91,7 @@ export const ToPatternLibrary = () => {
           </div>
         </axa-commercial-hero-banner>
         <img
-          class="landingpage__commercial-hero-banner-image"
+          class="welcome__commercial-hero-banner-image"
           src="axa-design-system-illustration.png"
           alt="Design system illustration"
         /><img />
@@ -83,4 +106,59 @@ export const ToPatternLibrary = () => {
 
 ToPatternLibrary.story = {
   name: 'to Pattern Library',
+};
+
+export const WhatIsNew = () => {
+  const wrapper = document.createElement('div');
+  wrapper.classList.add('accessory-story-wrapper');
+
+  const xhttp = new XMLHttpRequest();
+  // eslint-disable-next-line func-names
+  xhttp.onreadystatechange = function() {
+    if (this.readyState === 4 && this.status === 200) {
+      const responseJson = JSON.parse(xhttp.responseText);
+      document.querySelector(
+        '#welcome__github-response'
+      ).innerHTML = getFormattedGitCommitMessage(responseJson);
+      document.querySelector(
+        '#welcome__github-response-date'
+      ).innerHTML = getDateFromGitCommit(responseJson);
+    }
+  };
+  xhttp.open(
+    'GET',
+    'https://api.github.com/search/commits?q=repo:axa-ch/patterns-library+Publish&sort=author-date&order=desc',
+    true
+  );
+  xhttp.setRequestHeader('Accept', 'application/vnd.github.cloak-preview');
+  xhttp.send();
+
+  const template = html`
+    <style>
+      ${styles}
+    </style>
+    <div class="accessory-story-content">
+      <axa-heading rank="1" variant="secondary">What's new</axa-heading>
+      <axa-text variant="size-2">
+        Here you will find the latest updates on AXA Design System
+        development.
+      </axa-text>
+      <axa-heading rank="2" variant="secondary">Update log</axa-heading>
+      <axa-heading rank="5">Component versioning is now live!</axa-heading>
+      <p class="welcome__text-with-link">
+        We are happy to announce that we now support component versioning!
+        Our solution automatically injects package.json version information into component code at build time. Additionally, by making components version-aware at runtime, we can dynamically name custom element to include version information.
+        For more information check out
+        <axa-link href="https://github.com/axa-ch/patterns-library/blob/develop/COMPONENT_VERSIONING.md">Introduction</axa-link>.
+      </p>
+      
+      <axa-heading rank="2" variant="secondary">Last releases</axa-heading>
+      <axa-heading rank="6" class="welcome__response-date"><span id="welcome__github-response-date"></axa-heading>
+      <axa-text variant="size-2"><span id="welcome__github-response"></span></axa-text>
+    </div>
+    <pl-contact-footer></pl-contact-footer>
+  `;
+
+  render(template, wrapper);
+  return wrapper;
 };
