--- src/other/pages/what-is-new/index.scss
@@ -1,20 +1,0 @@
-@import '../utils/accessory-story.scss';
-@import '../utils/canvas-content-padding.scss';
-
-@include canvas-padding;
-@include content-padding;
-
-@include accessory-story-content;
-@include accessory-story-wrapper;
-
-.what-is-new {
-    &__text-with-link {
-        @include text(t2);
-        max-width: 750px;
-    }
-
-    &__response-date {
-        // The important flag is needed, because the component iself sets its styles to the host element internally.
-        margin-bottom: 20px !important;
-    }
-}
